complete rewrite
The code now conforms to the MediaWiki extension standard of one small file which is always included, one bigger file with the rest of the code which is only loaded when needed, an i18n file and a magic file for the parser function.
This extension now uses the {{#filelist:}} tag instead of <filelist />, wiki >> html.
The javascript and css are now loaded using the mediawiki standard way of doing this instead of inserting it directly into the body.. (TODO: fix the location of the $wgOut->addResources( ... ) call)